Bug #1727831 unisoc: sprdwcn: pclint scan slove int change unsigned long
[root cause  ] int type chang error
[changes     ] int change to unsigned long
[side effects]na
[self test   ]:yes
[download normally]:yes
[power on/off normally]:yes
[do common repository/branch inspection]:no
[is there dependence]:no
[confirm dependent commit]:no
[board]:ums512
[test case]:pclint
[reviewers   ] congcong.wang
[change_type ] feature_bugfix
[tag_product ] common

Test: int change to unsigned long

+++ drivers/unisoc_platform/sprdwcn/boot/wcn_integrate_boot.c
@@ -3105,7 +3105,7 @@ int start_integrate_wcn_truely(u32 subsys)
 	bool is_marlin;
 	struct wcn_device *wcn_dev;
 	u32 subsys_bit = 1 << subsys;
-	int ret_wait_completion = 0;
+	unsigned long ret_wait_completion = 0;
 
 	WCN_INFO("start subsys:%d\n", subsys);
 	wcn_dev = wcn_get_dev_by_type(subsys_bit);
